+++ app/models/api/token.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Api
+  class Token
+    extend HttpMethod
+
+    class << self
+      def create(email:, password:)
+        response = post(
+          '/token', params:  { email:, password: },
+                    headers: { 'Content-Type' => 'application/json' }
+        )
+        JSON.parse(response.body, symbolize_names: true)
+      end
+    end
+  end
+end
