+++ spec/models/api/token_spec.rb
@@ -2,9 +2,9 @@
 
 require 'rails_helper'
 
-RSpec.describe UsersApi do
-  describe '#create_token' do
-    subject { UsersApi.create_token(email:, password:) }
+RSpec.describe 'Token' do
+  describe '#create' do
+    subject { Api::Token.create(email:, password:) }
 
     context '管理者ユーザの場合' do
       let(:email) { Rails.application.credentials.app.rails_sample_email }
